Exercise 2.5 solution
