+++ courseinfo/src/App.js
@@ -1,45 +1,17 @@
 
-const Header = ({name}) => {
-    return (
-        <h2>{name}</h2>
-    )
-}
-
-const Content = ({parts}) => {
-    return (
-        parts.map(part =>
-            <Part key={part.id} part={part}/>
-        )
-    )
-}
-
-const Part = ({part}) => {
-    return (
-        <p>{part.name} {part.exercises}</p>
-    )
-}
-
-const Stats = ({parts}) => {
-    return (
-        <strong><p>total of {parts.reduce((sum, part) => {return sum + part.exercises}, 0)} exercises</p></strong>
-    )
-}
-
-const Course = ({course}) => {
-    return (
-        <>
-            <Header name={course.name} level="2"/>
-            <Content parts={course.parts}/>
-            <Stats parts={course.parts}/>
-        </>
-    )
-}
+import Course from './components/Course'
 
 const Courses = ({courses}) => {
     return (
         courses.map(course =>
             <Course key={course.id} course={course} />
         )
+    )
+}
+
+const Header = ({text}) => {
+    return (
+        <h1>{text}</h1>
     )
 }
 
@@ -91,11 +63,10 @@ const App = () => {
 
     return (
         <>
-            <h1>Web development curriculum</h1>
+            <Header text="Web development curriculum"/>
             <Courses courses={courses}/>
         </>
     )
-
 }
 
 export default App
